--- apache-beam/ctr_prediction/.gitignore
@@ -1,5 +1,0 @@
-.DS_Store
-.idea
-*.iml
-build
-target
